project updated

+++ ex03/bsp.cpp
@@ -16,11 +16,11 @@ static float area(Point const a, Point const b, Point const c)
 {
 	float area;
 
-	area =	(
-			(a.getX().toFloat() * (b.getY().toFloat() - c.getY().toFloat()))
-			+ (b.getX().toFloat() * (c.getY().toFloat() - a.getY().toFloat()))
-			+ (c.getX().toFloat() * (a.getY().toFloat() - b.getY().toFloat()))
-			) / 2;
+	area =	((
+			(a.getX() * (b.getY() - c.getY()))
+			+ (b.getX() * (c.getY() - a.getY()))
+			+ (c.getX() * (a.getY() - b.getY()))
+			) / 2).toFloat();
 	return((area >= 0) ? area : -area);
 }
